requirements.txt
# Core dependencies
google-generativeai>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Optional AI providers (uncomment as needed)
# openai>=1.0.0
# anthropic>=0.34.0
# langchain>=0.1.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0